requirements.txt
marionette-tg
python-nmap
requests
pyopenssl
pyasn1
